PagePublishToggle: notify on collection publish failure
Lifts CortextSnackbars to the route shell so notices reach every screen,
not just Canvas. Wraps the cascaded collection publish in throwOnError
+ Promise.allSettled so a single rejection surfaces a single snackbar
and the page itself stays unpublished.
