+++ src/components/Canvas.js
@@ -7,8 +7,7 @@ import {
 	InterfaceSkeleton,
 	store as interfaceStore,
 } from '@wordpress/interface';
-import { Button, Disabled, SnackbarList } from '@wordpress/components';
-import { store as noticesStore } from '@wordpress/notices';
+import { Button, Disabled } from '@wordpress/components';
 import { chevronDown, chevronUp, cog, seen, unseen } from '@wordpress/icons';
 import { useCallback, useEffect, useState } from '@wordpress/element';
 
@@ -37,28 +36,6 @@ import PageInspectorSidebar, {
 	PAGE_INSPECTOR,
 	isInspectorArea,
 } from './PageInspectorSidebar';
-
-// Renders only Cortext-owned snackbars (the autosave failure toast). The
-// editor store also dispatches its own "Post updated" success notice on every
-// save; in an autosave-silent UI those would fire constantly, so we filter to
-// notices we tagged ourselves.
-function CortextSnackbars() {
-	const notices = useSelect(
-		( select ) =>
-			select( noticesStore )
-				.getNotices()
-				.filter(
-					( n ) =>
-						n.type === 'snackbar' &&
-						typeof n.id === 'string' &&
-						n.id.startsWith( 'cortext-' )
-				),
-		[]
-	);
-	const { removeNotice } = useDispatch( noticesStore );
-
-	return <SnackbarList notices={ notices } onRemove={ removeNotice } />;
-}
 
 function DocumentActions( {
 	isActive,
@@ -292,7 +269,6 @@ function CanvasEditor( {
 				arePropertiesVisible={ arePropertiesVisible }
 				onTogglePropertiesVisible={ togglePropertiesVisible }
 			/>
-			<CortextSnackbars />
 			<HideHeaderBlockKebab />
 			<InterfaceSkeleton
 				className="cortext-canvas"
